Make CentOS a build-time option again
Big picture...we are going to cross a need for a freely
available version.  Otherwise too many things get painful;
it's not easy for other people to contribute.

So let's add CentOS back as an option.  I split the host
build in two.  I also deleted the rhel repo as only RHT
employees can use it today.  And in general people will
want to configure it anyways if they're doing custom
builds.