Add Jenkinsfile compose pipeline
Add a Jenkinsfile that runs the treecompose and publishes it. Right now,
this is only running the `rhcos` variant, so it's publishing to some
place private. The next steps are to also build cloud images, and to
also build the `origin` CentOS-based variant and push that out somewhere
public.

I brought back the RHEL repos here for simplicity; I'd like to keep as
much of the logic as possible right here. Though I'm open to changing
this back.