Use Makefile to separate responsibilities a bit more
Let's standardize on the `Makefile` as the CLI-friendly approach for
local development. E.g. `make syntax-check` to verify that the JSON
files are sane, and `make container` to build an update-in-container.
OTOH, `.prow.sh` still does `yum -y install jq`.

My goal is to then add more Makefile targets that can be used by humans
as well as by various CIs.