treecompose: Use $WORKSPACE
I'm not sure what's going on here, but the same thing we hit with the
rdgo job is happening with the treecompose. I think it might have to do
with the fact that pipelines can have workspaces across e.g. multiple
places. Anyway, this fixed it for rdgo, so let's try it.