treecompose: Add db-diff, and do write ref again
When I added --write-commitid-to we stopped setting the ref; fix
that by doing this in a different way.

Use this to add `db diff`.

I found this useful for FAHC.  Also...Groovy quoting vs shell
quoting is a huge mess.  I found
http://mrhaki.blogspot.com/2009/08/groovy-goodness-string-strings-strings.html
https://gist.github.com/Faheetah/e11bd0315c34ed32e681616e41279ef4
etc. very useful.

We should probably change our pipelines to use `chdir` and avoid
using source directories as strings.  Or maybe inject them
into the real environment?  Basically it feels saner to me if
we always use `sh('command')` and not `sh("command")`.