cloud: add metadata generation stage
The bulk of this commit adds a metadata generation stage to the cloud
pipeline.  Currently, it will generate the following:

  - complete package list for the ostree commit
  - package diff between previous ostree commit and HEAD
  - sha256sum of gzipped qcow2 image

These files land in the same directory as the gzipped qcow2 image.

In order to generate the package data, we have to rsync the ostree
repo as part of the 'Sync In' stage.