cloud: Check for no-op
Don't bother running imgfac if there aren't any new commits. As
mentioned in the comments, this is not strictly correct since there are
other input files that could have changed. In practice though, new
treecomposes are so much more frequent than changes in those other
inputs to make this a non-issue.