Add a "rollup" task that does ostreerepo → container
This repository's initial `Dockerfile` runs rpm-ostree inside
a container, generating a container as a result.  But to generate e.g.
`qcow2` with the target content (as opposed to bootstrap), we also
today are doing a path where we have a "traditional" split out repo.

This closes the loop and generates a container that just "rolls up"
the existing ostree repo - so we'll have the same commits in both.