Define parameters in Jenkins pipeline
Rather than requiring parameters to be set on the Jenkins job where the
pipeline is running, make it part of the definition here. This makes
their definitions here even more self-contained. To do this, I moved all
the internal strings into credentials that are then mapped to Groovy
variables when needed. (And just hardcoded a few less sensitive ones).

There's a gotcha I ran into early on, which is that on the first run,
parameters are not set yet so environment variables referencing them
would not work. I added a blurb about this in `pipeline-utils.groovy`.

We should also be able to also define the timer syntax using this
approach, though let's make sure this works smoothly first.