Set timer interval directly from pipeline
This moves the last bit of configuration from the internal JJB to the
pipeline. Another reason for doing this now is that I've noticed the
timer set by JJB is actually getting overwritten by the `properties`
call we do here. The end result is that the timers were completely
turned off.