container: Fix locations, tweak README.md
This broke when reworking the build system.  At some point we'll
have tests for this.  While here I cleaned up the naming a bit - I
like consistently having an OSTree repo named `repo`, and `oscontainer`
is way clearer than `local`.

And to reduce duplication, change the `index.html` to just talk about the
in-cluster service.