cloud: Fix vagrant box to use qcow2
I had only tested the pipeline ran, not the result as I'd been
doing a dry run.  We need a qcow2 inside the box.  Tweak things
so that we convert the raw to qcow2 first thing, and delete
the raw.  Then the VMDK conversion and vagrant box can both operate
on the qcow2, then gzip it when those are done.