cloud: Fixes for ec2 stamping
We were seeing errors doing the ec2 stamping; things seem
reliable for me when I do it from a qcow2, then convert that to vmdk.
I didn't entirely pin down the problem, but I suspect it's a bug
in the qemu VMDK layer - we're running a long unsupported and insecure
Fedora 25 OS for the cloud bits.  Will look at updating our stack
before debugging farther.

In the meantime, this seems to work for me, although I didn't
test the resulting AMI.