Fix ConditionFirstBoot= by removing /etc/machine-id
Requires: https://github.com/projectatomic/rpm-ostree/pull/1425

We want systemd's ConditionFirstBoot to fire. The
primary rationale here is that we're adopting Ignition for Fedora CoreOS,
and having ConditionFirstBoot= function will help a lot, as the idea
is it only runs once.  Further, this ensures that any presets generated
in the initramfs work.