jenkins/treecompose: Maintain content in container
Rather than having a "traditional" OSTree repository stored externally,
maintain it canonically inside a container image.  The "oscontainer/rollup"
task then goes away.

The advantages of this are that we now only have the OSTree repository
in exactly one place, and can e.g. consistently mirror it or promote it.

This PR prototypes out doing "podman in privileged containers"; I struggled
a bit with trying to use the Jenkins/docker APIs, and I think this ends
up being better.