Dockerfile: capture version + commit in image label
It would be useful to include the version and commit ID as part of the
image label.  I took a stab at using the `org.openshift.os` namespace
where the information could be stored.

The default values should be overwritten as long as the `build`
command passes values via `--build-arg`