oscontainer: actually store commit + version in labels
In #181, I changed a `Dockerfile` to include the commit ID and version
string in the label of the container image that would be generated.
What I failed to notice was that I made changes to a `Dockerfile` that
does not appear to be used in the pipeline at all.

This makes the same changes from that PR to `Dockerfile.rollup`, which
is actually used in the `treecompose` phase of the pipeline.