treecompose: explicitly set the ostree version number
The ostree repo only has a single commit in it and thus the version
string was not being changed with each change to the tree.  To move
away from that model, this change explicitly sets the `version`
metadata string as part of the ostree commit.

It's configured to use the previous `automatic_version_prefix` from
the treefile and appends the build number.