jenkins/cloud: Rework to use coreos-assembler's virt-install
Requires: https://github.com/cgwalters/coreos-assembler/pull/9

There are a lot of things going on here.  The *primary* goal though
is to drop the ImageFactory dependency because it's annoying to
containerize.  It fundamentally clashes with the way we were driving
it via Jenkins; our job would just find the first image in
`/var/lib/imagefactory`, meaning we couldn't do concurrent builds.

I spend my day-to-day development in a dev container, and while I
do leak in the libvirt socket, ImageFactory still has other problems
with that.

Also, this moves `coreos-oemid` to the assembler container, under
the philosophy that all *code* should live in that repo, and this
git repo just contains Jenkins orchestration + configs.

There were some unexpected hurdles to getting this to work:

 - When testing interactively, it went fine, but then I had to
   rework our console to log to a file for Jenkins with no tty
 - In Jenkins there were races/bugs due to libvirt (session)
   relying on the qemu process being reaped to determine shutdown.
   I tried to use `docker run --init` but ended up using bwrap
   in the container.
 - Kernel panics on the jenkins worker running Fedora 25/28 - I
   provisioned a RHEL7 worker and things went fine.

Overall though I think this is a lot better.