pipelines: skip the repo checkout
Jenkins will checkout the `openshift/os` repo into its workspace
before any of the stages are started.  The workspace is then mounted
into the container to be used by the stages.  We can skip checking out
the repo a second time, as it is automatically made available inside
the container.