jenkins: Add concept of "host workspace"
For treecompose right now, but I was planning to change the cloud
task to pull the container too, so let's add an API.

We need a "host workspace" because $WORKSPACE may be on overlayfs,
and one can't do overlayfs on overlayfs.

This solves a few problems.  First, we could easily "leak" content
across task invocations; say that a job created a file one time and
then stopped.  Cleaning is a good idea for the same reason workspace
cleanup is.

Second, hardcoded paths on the host clash badly with testing
the pipeline; we don't want my `cgwalters-test-pipeline` to potentially
stomp on data.