treecompose: Don't rsync pkgdiff to images
When we added multiple Jenkins nodes, this broke an implicit
assumption that the previous image content would be in the
previous single node's `/srv`, and hence caused treecompose
to delete the images if the job was scheduled on a node that
didn't have a cache.

Let's keep things simple and only have treecompose read/write
the oscontainer, and the cloud job only read/writes the `images/`.