Add a job to generate ootpa oscontainer
The treecompose job works, and one can pivot to it.  I have
more work to do to share code between the now 2 treecompose jobs.
And we're obviously missing cloud images (further patches to come
for that).

There are some SELinux denials and failing services (it looks
like preset is enabling sssd socket units somehow?)

But this is useful for us to add to the matrix and make available.