jenkins/cloud: Change version schema, add+use `meta.json`
This drops the date stamp, and switches to a version number
scheme `${ostree_version}-${genver}` where `genver` is like
`Release` in RPM terms.

The general principle here is that we make the ostree version
a unifying identifier across the oscontainer, vmimages etc.

Also, we now have more easily parsable JSON in the images directory.

Closes: https://github.com/openshift/os/issues/258