host: add coreos-metadata binary and service
This adds coreos-metadata RPM to RHCOS treefile. Current package is for
latest upstream version (v3.0.0) and provides the binary and a
service unit (disabled). The ssh-keys service is not relevant to RHCOS
and not installed in the rpm.

This consumes a dedicated copr repo:
https://copr.fedorainfracloud.org/coprs/lucab/rhcos-coreos-metadata/

Source specfile is at https://github.com/lucab/rhcos-coreos-metadata

Ref: COREOS-406