Readd ootpa pipeline using coreos-assembler
This took a lot of iteration.  The high level strategy is to
sync the previous build metadata from s3 (very cheap), which is
a major improvement over the previous image job.  We then sync out
the new build.

Per the coreos-assembler design, this job ties together the
treecompose + cloud images + oscontainer.  When we push an oscontainer,
it gets tagged with the build version - simple!

TODO:
 - GC for builds
 - Factor out a separate `fetch` trigger job
 - Tweak the cloud image to be pre-pivoted
 - Add basic kola sanity test
 - Add an aws-tested equivalent job (and symbolic tagging)

And of course the big task is to switch maipo over to this too.