jenkins: Add Slack notifications for job status changes
This wraps the body of most jobs in a try/catch block which sets
the run's status to failure if an exception or error is thrown.
At the end, that status is compared to the previous run, and a
notification is sent if they differ (treating anything other than
success as a failure).

The logic for sending notifications is stored in the utils file to
reduce code duplication, which means notifications will not be sent
for failures caused by changes to the utils file.

This expects that the Slack plugin is configured to be able to send
to the CoreOS Slack instance.  It only overrides the default Slack
channel, in case something else makes more sense for the global
configuration.