host: Switch to inline `postprocess` element
The primary reason I'm doing this is that today rpm-ostree's
treefile inheritance doesn't quite do what one wants with external
files.  There's some major work on that in
https://github.com/projectatomic/rpm-ostree/pull/1574
But let's not block on/require it.

Furthermore, using the inline `postprocess` value allows us
to clearly demarcate the `maipo`-specific hacks in that file.