aws-test: don't use `skopeo delete` to untag images
From containers/skopeo#531:

>  `skopeo delete` is not an “untag” operation: it deletes the image,
and all tags referring to it within the repository. It is not possible
to delete a single tag pointing at an image without deleting the image
and all other tags pointing at it at the same time; the registry API
does not provide such an operation.

Since the `delete` operation is going to fail, just remove it for now
to unblock the job until we get a workaround.