treecompose: use `oc` commands to tag images
This change forgoes the use of `skopeo copy` to tag the container
image with an ostree commit and instead directly talks to OpenShift to
apply the tag.  I think this should be less likely to fail as we
don't have to rely on a longer, more network intensive `skopeo copy`
operation.

Note: we are explicitly tagging the sha256sum reference of the image,
rather than re-tagging the `buildmaster` tag.  This means the ostree
commit tag will always point to the sha256sum of the image, rather
than the `buildmaster` tag, which changes each build.