40rhcos-fde: gate rootfs reprovisioning behind coreos.rootfs-replace.exp
Right now, with the latest FCOS, we get failures because e.g. the
rootmap code trips up on the dm-linear target trick we use.

This is a transitional state towards using the new rootfs reprovisioning
and LUKS-via-Ignition work. So for now so that we can keep bumping the
FCOS submodule, let's just gate all the reprovisioning services behind a
karg. That way, devs can still re-enable them if they'd like.