Document and standardize on rpm-repo names used by api.ci
Better document the fact that we don't include any `.repo`
files that you need to build this.

Update the manifest to use the repo names used by api.ci;
this *will break* current downstream RHCOS until we adjust
it to match these repo names.

I'll look at a patch for that.