overlay.d/05rhcos: add rhcos-upgrade-root-filesystem-container
Adds a new systemd service for the duration of 4.7 that runs at shutdown
to modify the partition table / kargs relating to the root device to
what they would be if the machine was freshly provisioned. This will
allow us to drop `coreos-cryptfs` entirely once we move to 4.8 because
we will no longer need to create dm-linear mappings for nocrypt root
devices to avoid performance hits & root LUKS devices will all be opened
by systemd / clevis in dracut.