overlay: nuke rhcos-growpart
Now that we've moved to Ignition LUKS, we don't need this anymore. It
also doesn't support multipath like `coreos-growpart` does.

(In fact, `ignition-ostree-growfs.service` has been running on RHCOS for
a while now because it was renamed from `coreos-growpart.service` and
moved to the initrd so the `disable` preset here didn't actually have
any effect.)