manifest: Re-add dhclient
With this, the dhclient binaries are added back to the RHCOS base.
Usage of dhclient as DHCP client in the initramfs remains unsupported.
In the real root, dhclient can be used as long as a configuration is provided.

In order to activate dhclient and use it instead of NetworkManager's internal DHCP client in the real root,
the user has to provide a configuration snippet via Ignition to configure NM to use dhclient,
e.g. in `/etc/NetworkManager/conf.d/dhclient.conf`:
```
[main]
dhcp=dhclient
```

Notably, the NetworkManager dispatcher script for dhclient.d config snippets remains excluded from the image.
This means if the admin desires to e.g. configure NTP via DHCP using dhclient,
a NM dispatcher script must be provided and placed at `/etc/NetworkManager/dispatcher.d/11-dhclient` ([1]),
so the `/etc/dhcp/dhclient.d/chrony.sh` script that is shipped with the chrony package is executed.

[1] https://src.fedoraproject.org/rpms/dhcp/blob/master/f/11-dhclient

RHBZ: https://bugzilla.redhat.com/show_bug.cgi?id=1908462