chrony NM dispatcher: Account for dhclient dispatcher script in /usr/lib/
In current dhcp-client Fedora RPMs,
the dhclient dispatcher script is placed at
`/usr/lib/NetworkManager/dispatcher.d/11-dhclient`
as opposed to
`/etc/NetworkManager/dispatcher.d/11-dhclient`
used before.

This changes the chrony dispatcher script to account for the
dhclient dispatcher script in either place.

This isn't strictly needed for RHCOS, but we want to keep
the script in sync with the one shipped in the chrony RPM:
https://src.fedoraproject.org/rpms/chrony/pull-request/5