Add new extension qemu-kiwi
This adds qemu-kiwi as another extension to the list. qemu-kiwi and its
dependencies are in the rhel-8-advanced-virt repository, so we have to
add this as well.

The qemu-kiwi extension is added to MCO with this PR
https://github.com/openshift/machine-config-operator/pull/2376

qemu-kiwi is a shrunk-down version of the qemu-kvm package with fewer
dependencies. 8 rpms and ~20MB in size when installed