ci: Rework to split build/test
Aiming to fix https://github.com/openshift/os/issues/506

I think the problem here may be related to the fact that
we're trying to do all of our CI as part of a build.

Instead, let's function like e.g. the CI for coreos/rpm-ostree
which creates a derived container from the source code.  For
us, this derived container is basically the 2-tuple (source, cosa).

For coreos/rpm-ostree we actually do generate binaries as well
as test code from the buildroot, i.e. a 3 tuple (binaries, test binaries, cosa).
But here we don't need that, there's no actual "build" step.

Split off `validate.sh` as a separate thing that can be a
clearly separate step.