Add sandboxed-containers extension
This reworks af93514887350da53e035d1b841eea36542cdd42 as, since it was
merged, a few changes happened in the enhancement proposal and it's been
decided to go with a "sandboxed-containers" extensions instead.

As the "qemu-kiwi" extension is no longer needed, we're removing it
(well, reworking it) and adding the "sandboxed-containers" one instead.

The sandboxed-containers extension is added to the MCO with by the
following PR:
https://github.com/openshift/machine-config-operator/pull/2456

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>