tests: Add checks for CLHM presets
In older versions of console-login-helper-messages (< v0.21) which
RHCOS is still on, issuegen-related systemd units exist and need
to be enabled for CLHM to function properly. Test that we have
those presets enabled in RHCOS if we're still on a version less
than v0.21. This test can be removed after RHCOS gets v0.21 or later.