extensions: Add oci-seccomp-bpf-hook
The oci-seccomp-bpf-hook package enables recording of secomp profiles,
which, apart from being useful on its own, is a part of the Security
Profiles Operator:
    https://github.com/openshift/enhancements/pull/745

This package was previously evaluated for inclusion into FCOS:
    https://github.com/coreos/fedora-coreos-tracker/issues/887
and it was recommended that we try adding the hook as an extension
instead.

Note that bcc (a dependency of oci-seccomp-bpf-hook) seems to have a weak
dependency on bcc-tools which is not needed for this case. I'm unsure
if there is a way to tell the extensions system to not install the weak
dependency though.