Use `deploy-via-container`
This will cause us to run through the ostree-native container
stack when generating the disk images.

Today for RHCOS we're using the "custom origin" stuff which
lets us inject metadata about the built source, but rpm-ostree
doesn't understand it.

With this in the future (particularly after https://github.com/coreos/coreos-assembler/issues/2685)
`rpm-ostree status` will show the booted container and *understand it*.

We'll have the digest of the OCI archive at least...though
that may get changed if it gets converted to docker v2s2 when pushing
to a registry...

Now in the future what we want is to entirely rework our build
pipeline like this: https://github.com/coreos/coreos-assembler/issues/2685