Bug 2017471: ensure nss-altfiles comes from AppStream
In #637, it was observed that `nss-altfiles` could come from the RHAOS
repos, which is unnecessary since the same version exists in the RHEL8
AppStream repos.  The package has since been untagged from the RHAOS
repo, but to ensure that we always get the package from AppStream we'll
pin to the specific version.  (The version hasn't changed since Apr 2019)