scripts/upload-ami: add support for making Aliyun images public
This allows us to configure Aliyun images as publicly available,
similar to what we do with AMIs.

I refactored a good chunk of the code to be a bit more reusable across
clouds.

Ideally, we'd have more of this in `coreos-assembler`, but this is
(hopefully) a short-term bandaid.

(See coreos/coreos-assembler#2604 where the `aliyun` CLI was added.)