scripts: add utility for marking Aliyun images public
Frustratingly, replicated copies of a public image in Aliyun do not
inherit the publicly available properties of the original image. So we
have to go back and mark the images as public after they have been
replicated.

I opted for a smaller utility script for this, rather than trying to
overload the already ballooning `upload-ami` script.

Requires openshift/os#688