Use new `overlay.d/cosa-no-autolayer` option
Tell coreos-assembler to not automatically add the overlays to the final
manifest. Add references to the auto-generated OSTree layers in our
manifest.

This isn't strictly necessary for RHCOS like it is for FCOS because the
branching model is simpler here. But doing this will allow us to not
support it both ways in cosa and make this the default behaviour so we
can drop the stamp file.

Requires: https://github.com/coreos/coreos-assembler/pull/2641