Use the generic redhat-release package
The redhat-release-coreos package is mostly a copy of the generic
package as all major configuration changes are made in this repo.

Thus move the os-release changes to a step in a postprocess script and
use the generic package to benefit from updates for free.

The changes are:

- No longer provide "extracted" GPG keys as they are not used in RHCOS
  (was in /usr/share/ostree/trusted.gpg.d/)
- containerisvsign.asc will be provided by the RHEL 8.5 version of this
  package in /etc/pki/rpm-gpg/
- Drop /etc/profile.d/path.sh
- Update default presets to their latest version as part of the
  redhat-release package

We do not add a VARIANT & VARIANT_ID qualifiers yet as this requires
changes in kola.