ci: fetch the 8.6 appstream repo for kata-containers
In order to keep successfully building the extensions, we need access
to the RHEL 8.6 Beta repo to fetch the `qemu-kvm-core` dependency that
`kata-containers` now requires.

This is a one-liner that fetches the 8.6 repos, greps for the
Appstream definition, and replaces the name/id. The output is written
to the existing `ocp.repo` file.