manifests: drop cryptsetup-reencrypt
The use of `cryptsetup-reencrypt` goes back to the null-cipher LUKS
container that we used to do disk encryption in early, early OCP 4
days.  As part of #788, the `rhcos-fde` dracut module dropped the
requirement on this package, so we should be able to safely drop it
from the manifest as well.