Switch to https://access.redhat.com/labs/rhir for bug reporting
OCP has migrated away from Bugzilla, so update the links
there.

I explicitly chose *not* to remove the
```
      REDHAT_BUGZILLA_PRODUCT="OpenShift Container Platform"
      REDHAT_BUGZILLA_PRODUCT_VERSION="${OCP_RELEASE}"
```

metadata because it's absolutely still valid to look there for
old bugs, and I am not aware of corresponding metadata keys for
JIRA - we could invent them of course, but it doesn't seem important
right now.