docs: Add rhaos-overrides.yaml
I think the source of truth for what we ship should be defined
via git and changes that are tested via pull requests; not humans
manually invoking a Koji tag command.

We don't yet live in that world and may never for RHEL-8, though
we have the opportunity with C9S to do better.

For now, this file defines what we intend to ship in a public
human (and in theory machine) readable way.