extensions: use Fedora as a builder
RHEL doesn't ship `createrepo_c`, so we were using
CentOS Stream 8 for that.

Then we hit on the fact that C8S doesn't exist for s390x, so switched
to C9S.

But RHEL9 made a push to require newer hardware baselines.  This
is problematic for parts of the RHCOS pipeline which are deployed
on Power8 hardware.

So...Fedora hasn't made the hardware baseline switches, so for
now let's just use that as a lowest common denominator.  The base
image and packages exist on all architectures we care about.