Add platform-specific console configuration metadata
cosa create_disk.sh and gf-set-platform will read this.  create_disk.sh
will also save a JSON version for the relevant architecture to
/boot/coreos/platforms.json for coreos-installer.

Enable graphical console on cloud platforms that allow capturing
screenshots, even if the graphical console is not interactive.

This is a copy of https://github.com/coreos/fedora-coreos-config/pull/1781.
Once the config has been ratcheted into testing-devel, we'll replace this
file with a symlink.