rhel9: Add `/etc/localtime` pointing to `UTC` by default
See https://issues.redhat.com/browse/LOG-3117

Something is creating it for RHCOS8, but seemingly not in the RHCOS9
build.  (I bet Anaconda does this by default, so it will be on
most Anaconda-based systems though)

This ensures that containers can mount it from the host; though
we should also chase down and fix those containers to do this in
a better way (one approach is to mount all of `/etc`).