platforms: add secondary serial console on OpenStack/VirtualBox/VMware
Previously, `openstack console log <server>` was useful for collecting
console logs from OpenStack VMs, and the serial console was similarly
useful in VMware, but we disabled serial console output on both
platforms.  Re-enable it as a secondary console so it gets as much
information as possible without interfering with the graphical console.
Do the same on VirtualBox to stay in sync with FCOS.

Fixes: https://issues.redhat.com/browse/OCPBUGS-2926
xref: https://github.com/coreos/fedora-coreos-config/pull/2038